+++ Assets/deCrawl/Scrips/Primitives/StateSaver.cs
@@ -1,0 +1,9 @@
+namespace DeCrawl.Primitives
+{
+    public interface StateSaver
+    {
+        public string SerializeState();
+
+        public void DeserializeState(string json);
+    }
+}
